requirements.txt
asgiref==3.7.2
certifi==2023.7.22
channels==4.0.0
charset-normalizer==3.2.0
Django==4.2.5
djangorestframework==3.14.0
future==0.18.3
idna==3.4
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tk==0.1.0
typing_extensions==4.8.0
urllib3==2.0.5
